Make basic wppl model run
Basic model has no subject-wise variables.
Still needs to be printed to csv in the right format.

Linking function for the binned-histogram data is a dirichlet.
Hence, participants who put the slider all the way to the endpoints (notably, 0),
had their responses replace with the next lowest slider value (0.01).
This is what is found is bin_data_noZeros.csv

Also added a timing function to priorUtils.js